requirements.txt
yfinance
investpy
pandas
requests
python-dotenv
